Improve Contact Card on Home
