+++ lib/base/localization/app_de.arb
@@ -21,7 +21,8 @@
    "tokenPermissions":"Berechtigungen für Token",
    "permissionChangePossibleInTUMonline":"Du kannst deine Berechtigungen in TUMOnline ändern",
    "login":"Anmelden",
-   "notLoggedIn":"Nicht angemeldet",
+   "notLoggedIn":"Nicht angemeldet!",
+   "profileError":"Profil konnte nicht geladen werden!",
    "welcomeToTheApp":"Willkommen in der\n TUM Campus App",
    "enterYourIDToStart":"Gib deine TUM ID ein, um zu starten",
    "continueWithoutID":"Ohne TUM ID fortfahren",
