+++ lib/base/localization/app_en.arb
@@ -21,7 +21,8 @@
    "tokenPermissions":"Token Permissions",
    "permissionChangePossibleInTUMonline":"You can change your permissions on TUMOnline",
    "login":"Login",
-   "notLoggedIn":"Not Logged In",
+   "notLoggedIn":"Not Logged In!",
+   "profileError":"Error fetching profile!",
    "welcomeToTheApp":"Welcome to the TUM Campus App",
    "enterYourIDToStart":"Enter your TUM ID to get started",
    "continueWithoutID":"Continue without TUM ID",
