+++ lib/homeComponent/contactComponent/views/contact_card_error_view.dart
@@ -1,0 +1,31 @@
+import 'package:campus_flutter/base/extensions/context.dart';
+import 'package:flutter/material.dart';
+
+class ContactCardErrorView extends StatelessWidget {
+  const ContactCardErrorView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(10.0),
+      child: Row(
+        children: [
+          const CircleAvatar(
+            backgroundImage: AssetImage(
+              'assets/images/placeholders/portrait_placeholder.png',
+            ),
+            radius: 50,
+          ),
+          const Padding(padding: EdgeInsets.only(left: 15)),
+          Expanded(
+            child: Text(
+              context.localizations.profileError,
+              style: Theme.of(context).textTheme.titleLarge,
+              textAlign: TextAlign.center,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
